- InputMap and OutputMap, I've realized, are pages (the directory formerly known as 'views'), so I've moved them into ./pages alongside such heavy hitters as Login.jsx and HomeScreen.jsx and HomeScreen II: The Sequel.jsx ;)
