+++ client/src/App.js
@@ -10,8 +10,8 @@ import {
 // Containers/high-level interfaces
 import HomeScreen from './pages/HomeScreen';
 import MainNavigation from './shared/components/Navigation/MainNavigation';
-import InputMap from './shared/components/UIElements/InputMap';
-import OutputMap from './shared/components/UIElements/OutputMap';
+import InputMap from './pages/InputMap';
+import OutputMap from './pages/OutputMap';
 import Login from './pages/Login';
 
 // Helpers
