+++ client/src/pages/InputMap.css
